implementing mapping many to one and one to many

+++ JavaGuruChat/src/com/seventysevenagency/chat/dao/MessageDAO.java
@@ -7,7 +7,7 @@ import com.seventysevenagency.chat.domain.Message;
 import com.seventysevenagency.chat.domain.User;
 
 public interface MessageDAO {
-	public Long create(Message message) throws DAOException;
+	public int create(Message message) throws DAOException;
 
 	public void update(Message message) throws DAOException;
