+++ JavaGuruChat/src/com/seventysevenagency/chat/dao/UserDAO.java
@@ -7,9 +7,9 @@ public interface UserDAO {
 
 	public void update(User user) throws DAOException;
 
-	public User findById(long id) throws DAOException;
+	public User findById(int id) throws DAOException;
 
-	public int deleteById(long id) throws DAOException;
+	public int deleteById(int id) throws DAOException;
 	
 	public User findByUsername(String username) throws DAOException;
 }
